Refactoring, Autoformatting tests to a separate module
